requirements.txt
Django>=4.2.16
psycopg2-binary>=2.9.0
boto3>=1.26.0
botocore>=1.29.0
python-dotenv>=1.0.0
Twisted>=24.7.0
autobahn>=22.7.1
service-identity>=21.1.0
cryptography>=43.0.0
attrs>=21.4.0